requirements.txt
langchain-openai
python-dotenv
fastapi
uvicorn
EOF